pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bsort"
version = "0.1.0"
description = "Real-time bottle cap detection and color classification optimized for edge devices (RPi 5)"
readme = "README.md"
requires-python = ">=3.10"
# Fixed: Changed from table {text="MIT"} to string to satisfy new Setuptools requirements
license = "MIT"
authors = [
    {name = "Candidate Name", email = "candidate@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "ultralytics>=8.0.0",        # YOLOv8
    "opencv-python-headless>=4.8.0", # Headless for Docker/CI stability
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "wandb>=0.15.0",             # Experiment tracking
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
    "click>=8.1.0",              # CLI framework
    "onnxruntime>=1.16.0",       # REQUIRED for <10ms inference
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "pylint>=2.17.0",
    "mypy>=1.4.0",
    "types-PyYAML",
]

[project.scripts]
bsort = "bsort.cli:main"

# CRITICAL FIX: Explicitly tell setuptools where the code is
# This prevents it from trying to install 'data' and 'configs' as python packages
[tool.setuptools]
packages = ["bsort"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pylint.master]
ignore = ["CVS"]
extension-pkg-allow-list = ["cv2", "torch", "onnxruntime"]

[tool.pylint.messages_control]
max-line-length = 100
disable = [
    "C0114", # Missing module docstring
    "C0115", # Missing class docstring
    "C0116", # Missing function docstring
    "R0903", # Too few public methods
    "R0913", # Too many arguments (common in ML configs)
    "E1101", # No member (common with dynamic libraries like cv2/torch)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=bsort --cov-report=term-missing --cov-report=xml"